requirements.txt
customtkinter
discord.py
openai
python-dotenv